+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Version 0.1.4
 
   * Osm::Model has new class method get\_user\_permission(api, section_id, permission)
+  * API may return permissions value as a string not integer
 
 ## Version 0.1.3